Timezone/Time Stamp Challenge
Puzzle #2: Dates and Timezones

Twitter has users around the world.  We need to be able to represent time in
our databases so that when we display the time of a tweet, no matter where you
are on the planet you see the time in the right timezone.  The safest way we do
this is to store time in milliseconds from the epoch.  When we retrieve the time,
we convert it to the proper display timezone.

Used Java libraries (TimeZone) to solve the coding problem 
https://docs.oracle.com/javase/7/docs/api/